pyproject.toml
[tool.poetry]
name = "cars"
version = "0.1.0"
description = ""
authors = ["odai1990 <odai22odai78@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.7.2"
whitenoise = "^5.3.0"
django-environ = "^0.4.5"
psycopg2-binary = "^2.9.1"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
